Added clean_dir method to utils and reused. Added constant.py for storing constants.
