+++ constants.py
@@ -1,0 +1,1 @@
+EXTENSION_TO_SKIP = [".png",".jpg",".jpeg",".gif",".bmp",".svg",".ico",".tif",".tiff"]
