+++ utils.py
@@ -1,0 +1,14 @@
+import os
+from constants import EXTENSION_TO_SKIP
+
+def clean_dir(directory):
+    # Check if the directory exists
+    if os.path.exists(directory):
+        # If it does, iterate over all files and directories
+        for dirpath, _, filenames in os.walk(directory):
+            for filename in filenames:
+                _, extension = os.path.splitext(filename)
+                if extension not in EXTENSION_TO_SKIP:
+                    os.remove(os.path.join(dirpath, filename))
+    else:
+        os.makedirs(directory, exist_ok=True)
